Reversed col and row order to match backgroundGrid
It turns out the background grid needs to referenced by row then column in order to make maps easier to read.  This is opposite for a lot of the draw code.  I went into the al the for loops and reversed them.  Also changed arrays and some functions as well.  I'm hoping this caught majority of the bugs.